requirements.txt
Flask
Flask-MySQLdb
python-dotenv
groq
Werkzeug
mysql-connector-python
gunicorn